// src/enums/param-type.enum.ts
export enum ParamType {
  // String types
  UUID = 'uuid',
  FullName = 'fullName',
  FirstName = 'firstName',
  LastName = 'lastName',
  Email = 'email',
  Password = 'password',
  PhoneNumber = 'phoneNumber',
  Sentence = 'sentence',

  // Boolean Types
  Boolean = 'boolean',
  true = 'true',
  false = 'false',

  // Numeric types
  int = 'int',
  float = 'float',
}